+++ Source/Shaders/CentralBodyVS.glsl
@@ -37,20 +37,29 @@ vec4 getPosition2DGeographicMode(vec3 position3DWC)
 
 vec4 getPosition2DWebMercatorMode(vec3 position3DWC)
 {
-    // TODO: only do this transformation for low LODs.
-    float currentLatitude = mix(u_southLatitude, u_northLatitude, textureCoordinates.y);
-    float sinLatitude = sin(currentLatitude);
-    float mercatorY = 0.5 * log((1.0 + sinLatitude) / (1.0 - sinLatitude));
-
-    // mercatorY - u_southMercatorY in simulated double precision.
-    float t1 = 0.0 - u_southMercatorYLow;
-    float e = t1 - 0.0;
-    float t2 = ((-u_southMercatorYLow - e) + (0.0 - (t1 - e))) + mercatorY - u_southMercatorYHigh;
-    float highDifference = t1 + t2;
-    float lowDifference = t2 - (highDifference - t1);
-    
-    float mercatorFraction = highDifference * u_oneOverMercatorHeight + lowDifference * u_oneOverMercatorHeight;
-    
+    // The width of a tile, in radians, at level 11 (assuming a single root tile) is
+    //   2.0 * czm_pi / pow(2.0, 11.0)
+    // We want to just linearly interpolate the 2D position from the texture coordinates
+    // when we're at this level or higher.  The constant below the expression
+    // above evaluated and then rounded up at the 4th significant digit.
+    const float maxTileWidth = 0.003068;
+    float mercatorFraction = textureCoordinates.y;
+    if (u_northLatitude - u_southLatitude > maxTileWidth)
+    {
+	    float currentLatitude = mix(u_southLatitude, u_northLatitude, textureCoordinates.y);
+	    float sinLatitude = sin(currentLatitude);
+	    float mercatorY = 0.5 * log((1.0 + sinLatitude) / (1.0 - sinLatitude));
+	
+	    // mercatorY - u_southMercatorY in simulated double precision.
+	    float t1 = 0.0 - u_southMercatorYLow;
+	    float e = t1 - 0.0;
+	    float t2 = ((-u_southMercatorYLow - e) + (0.0 - (t1 - e))) + mercatorY - u_southMercatorYHigh;
+	    float highDifference = t1 + t2;
+	    float lowDifference = t2 - (highDifference - t1);
+	    
+	    mercatorFraction = highDifference * u_oneOverMercatorHeight + lowDifference * u_oneOverMercatorHeight;
+    }
+        
     vec4 rtcPosition2D = vec4(0.0, mix(u_tileExtent.st, u_tileExtent.pq, vec2(textureCoordinates.x, mercatorFraction)), 1.0);  
     return u_modifiedModelViewProjection * rtcPosition2D;
 }